+++ spec/models/service_spec.rb
@@ -37,4 +37,16 @@ RSpec.describe Service do
 
     expect(service.main_category).to eq(main)
   end
+
+  context "if open access" do
+    before { allow(subject).to receive(:open_access?) { true } }
+
+    it { is_expected.to validate_presence_of(:connected_url) }
+  end
+
+  context "if not open access" do
+    before { allow(subject).to receive(:open_access?) { false } }
+
+    it { is_expected.to_not validate_presence_of(:connected_url) }
+  end
 end